fix : added double in sell order

+++ models/sell.js
@@ -1,39 +1,37 @@
-
-module.exports = (sequelize,DataTypes) =>{
+module.exports = (sequelize, DataTypes) => {
     const Sell = sequelize.define(
-        "sell",
-        {
+        "sell", {
             id: {
-                type : DataTypes.INTEGER,
-                allowNull:false,
+                type: DataTypes.INTEGER,
+                allowNull: false,
                 primaryKey: true,
                 autoIncrement: true,
-                validate:{
-                    notEmpty:true,
+                validate: {
+                    notEmpty: true,
                 }
             },
             qty: {
-                type : DataTypes.INTEGER,
-                allowNull:false,
-                validate:{
-                    notEmpty:true,
+                type: DataTypes.DOUBLE,
+                allowNull: false,
+                validate: {
+                    notEmpty: true,
                 }
             },
             price: {
-                type : DataTypes.INTEGER,
-                allowNull:false,
-                validate:{
-                    notEmpty:true,
+                type: DataTypes.DOUBLE,
+                allowNull: false,
+                validate: {
+                    notEmpty: true,
                 }
             },
             total_price: {
-                type : DataTypes.INTEGER,
-                allowNull:false,
-                validate:{
-                    notEmpty:true,
+                type: DataTypes.DOUBLE,
+                allowNull: false,
+                validate: {
+                    notEmpty: true,
                 }
             }
         }
-    ); 
+    );
     return Sell;
 };
